requirements.txt
fastmcp==2.13.0.2
weaviate-client>=4.17.0,<5
starlette>=0.37
